+++ packages/imperative/src/config/src/ProfileInfo.ts
@@ -842,6 +842,25 @@ export class ProfileInfo {
         this.loadAllSchemas();
     }
 
+    //_________________________________________________________________________
+    /**
+     * Function to ensure the credential manager will load successfully
+     * Returns true if it will load, or the credentials are not secured. Returns false if it will not load.
+     */
+    public async profileManagerWillLoad(): Promise<boolean> {
+        if (this.mCredentials.isSecured) {
+            try {
+                await this.mCredentials.loadManager();
+                return true;
+            } catch (err) {
+                this.mImpLogger.warn("Failed to initialize secure credential manager: " + err.message);
+                return false;
+            }
+        } else {
+            return true;
+        }
+    }
+
     /**
      * Returns whether a valid schema was found (works for v1 and v2 configs)
      */